Improve Item properties tag writing logic
This makes how metadata changes are saved in Item properties consistent with how they are saved in the playlist view.

This includes sharing more logic and pushing more logic to the `file_info_filter` implementation.
